+++ internal/adv/middlewares/authorize.go
@@ -2,6 +2,7 @@ package middlewares
 
 import (
 	"fmt"
+	"log"
 	"net/http"
 
 	advjwt "github.com/thongpham95/adv-tv-backend/internal/adv/utils/jwt"
@@ -18,13 +19,11 @@ type UserBody struct {
 func IsAuthorized(endpoint func(http.ResponseWriter, *http.Request)) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		if r.Header["Authorization"] == nil {
-			fmt.Println("Unauthorized")
 			responsehandler.NewHTTPResponse(false, "Unauthorized", nil).ErrorResponse(w, http.StatusUnauthorized)
 		} else {
-			fmt.Println("Getting token from header")
 			tokenStr, err := advjwt.ValidateToken(r.Header["Authorization"][0])
 			if err != nil {
-				fmt.Println("Err Getting token from header: ", err)
+				log.Println("Error getting token from header: " + err.Error())
 				responsehandler.NewHTTPResponse(false, err.Error(), nil).ErrorResponse(w, http.StatusUnauthorized)
 			}
 			if tokenStr != nil {