+++ internal/adv/models/media.go
@@ -4,10 +4,13 @@ package models
 type MediaItem struct {
 	ID       string `json:"id"`
 	DeviceID string `json:"device_id"`
+	IsVideo  bool   `json:"is_video"`
 	Key      string `json:"key"`
 }
 
 // MediaItemRepository exported
 type MediaItemRepository interface {
-	Create(deviceID string, mediaItem MediaItem) error
+	GetMediaBasedOnID(ID string) (*MediaItem, error)
+	GetMediaBasedOnKey(key string) (*MediaItem, error)
+	Create(deviceID string, mediaItem *MediaItem) error
 }