+++ CHANGELOG.md
@@ -21,7 +21,10 @@
 ## 1.1.2v - 03/08/2020
 - feat/upload-file:
 	+ Delete error handler, remove it dependencies.
-## 1.1.3v - 03/08/2020
+## 1.2.0v - 05/08/2020
 - feat/upload-file:
 	+ Add return statement after handling errors.
+	+ Add SignUp API.
+	+ Add hashing password into Login + SignUp API.
+	+ Add Add Device API + Get File URL + Upload file to device.